Cleanups of build.
Provide a somewhat-sane build.properties file, which has at least a
small chance of working out-of-box (versus the old build.properties,
which had zero chance of working).

Remove references to prior developer's home directory from GUMS client
and improve the logging on failure of VOMS (particularly, we now get
the exception logged when you try to call out to GUMS on RHEL7).