strip '-i' arg in sudoCommand for sudoTest (SOFTWARE-1721)
This is kind of a hack, but 'sudo -i' is supposed to be allowed, but the
sudo test command has '-v' (verify), which doesn't work with '-i' ...

Just using a hard coded 'sudo -v -S' would probably be fine (maybe even
better), but if, say, sudoCommand contained an alternate fully-qualified
path to a custom sudo, that would break.  Not sure if we care, but we'll
try to cover our bases anyway.